fix: fixed halves missing from series number in Hardcover metadata

+++ booklore-api/src/main/java/org/booklore/service/metadata/parser/hardcover/GraphQLResponse.java
@@ -252,7 +252,7 @@ public class GraphQLResponse {
     @AllArgsConstructor
     @JsonIgnoreProperties(ignoreUnknown = true)
     public static class FeaturedSeries {
-        private Integer position;
+        private Float position;
         private Series series;
     }
